+++ Zoop/Core/IZoopController.cs
@@ -1,4 +1,5 @@
 using Assets.Scripts.Inventory;
+using ZoopMod.Zoop.Preview;
 
 namespace ZoopMod.Zoop.Core;
